Deduplicate `struct Av1RestorationUnit` to `src/lf_mask.rs`.
